Initial version of columns block
This is a block that composes of multi column.
It starts with one, and the user can click on 
the new button on the UI to add more column.

Each column has the equal width.

Each column has a top-right button to remove 
itself. This has not been implemented, but 
the button is there.

CSS is inline because I don't want to create
separate CSS file yet.